+++ SampleAPIClass.class.php
@@ -1,0 +1,7 @@
+<?php
+
+class SampleAPIClass {
+    public function outputThreeParameters($param1, $param2, $param3) {
+        return "We received ($param1, $param2, $param3)";
+    }
+}